+++ app/src/main/java/net/maxproit/salesforce/model/myprospect/updatemyprospect/Data.java
@@ -14,10 +14,10 @@ public class Data {
 
     @SerializedName("agriculturalIncome")
     @Expose
-    private double agriculturalIncome;
+    private Integer agriculturalIncome;
     @SerializedName("apartmentIncome")
     @Expose
-    private double apartmentIncome;
+    private Integer apartmentIncome;
     @SerializedName("assetType")
     @Expose
     private String assetType;
@@ -32,13 +32,13 @@ public class Data {
     private String branchName;
     @SerializedName("businessIncome")
     @Expose
-    private double businessIncome;
+    private Integer businessIncome;
     @SerializedName("coApplicants")
     @Expose
     private List<CoApplicant> coApplicants = new ArrayList<>();
     @SerializedName("commercialSpaceIncome")
     @Expose
-    private double commercialSpaceIncome;
+    private Integer commercialSpaceIncome;
     @SerializedName("company")
     @Expose
     private String company;
@@ -75,13 +75,13 @@ public class Data {
     private String eTin;
     @SerializedName("emiOfOtherLoan")
     @Expose
-    private double emiOfOtherLoan;
+    private Integer emiOfOtherLoan;
     @SerializedName("factoryIncome")
     @Expose
-    private double factoryIncome;
+    private Integer factoryIncome;
     @SerializedName("familyExpenditure")
     @Expose
-    private double familyExpenditure;
+    private Integer familyExpenditure;
     @SerializedName("fatherName")
     @Expose
     private String fatherName;
@@ -93,7 +93,7 @@ public class Data {
     private Float fee;
     @SerializedName("interestIncomeOfFDR")
     @Expose
-    private double interestIncomeOfFDR;
+    private Integer interestIncomeOfFDR;
     @SerializedName("intersetRate")
     @Expose
     private Float intersetRate;
@@ -102,7 +102,7 @@ public class Data {
     private String leadReferenceNo;
     @SerializedName("loanRequired")
     @Expose
-    private double loanRequired;
+    private Integer loanRequired;
     @SerializedName("loanTerm")
     @Expose
     private Integer loanTerm;
@@ -132,7 +132,7 @@ public class Data {
     private String motherTitle;
     @SerializedName("netSalary")
     @Expose
-    private double netSalary;
+    private Integer netSalary;
     @SerializedName("netSalaryType")
     @Expose
     private String netSalaryType;
@@ -189,19 +189,19 @@ public class Data {
     private String relationshipWithApplicant;
     @SerializedName("remittanceIncome")
     @Expose
-    private double remittanceIncome;
+    private Integer remittanceIncome;
     @SerializedName("rmCode")
     @Expose
     private String rmCode;
     @SerializedName("securityValue")
     @Expose
-    private double securityValue;
+    private Integer securityValue;
     @SerializedName("segment")
     @Expose
     private String segment;
     @SerializedName("semipakaIncome")
     @Expose
-    private double semipakaIncome;
+    private Integer semipakaIncome;
     @SerializedName("spouseName")
     @Expose
     private String spouseName;
@@ -213,7 +213,7 @@ public class Data {
     private String status;
     @SerializedName("tutionIncome")
     @Expose
-    private double tutionIncome;
+    private Integer tutionIncome;
     @SerializedName("userName")
     @Expose
     private String userName;
@@ -251,19 +251,19 @@ public class Data {
     public void setRuleEngineInformation(String ruleEngineInformation) {
         this.ruleEngineInformation = ruleEngineInformation;
     }
-    public double getAgriculturalIncome() {
+    public Integer getAgriculturalIncome() {
         return agriculturalIncome;
     }
 
-    public void setAgriculturalIncome(double agriculturalIncome) {
+    public void setAgriculturalIncome(Integer agriculturalIncome) {
         this.agriculturalIncome = agriculturalIncome;
     }
 
-    public double getApartmentIncome() {
+    public Integer getApartmentIncome() {
         return apartmentIncome;
     }
 
-    public void setApartmentIncome(double apartmentIncome) {
+    public void setApartmentIncome(Integer apartmentIncome) {
         this.apartmentIncome = apartmentIncome;
     }
 
@@ -299,11 +299,11 @@ public class Data {
         this.branchName = branchName;
     }
 
-    public double getBusinessIncome() {
+    public Integer getBusinessIncome() {
         return businessIncome;
     }
 
-    public void setBusinessIncome(double businessIncome) {
+    public void setBusinessIncome(Integer businessIncome) {
         this.businessIncome = businessIncome;
     }
 
@@ -315,11 +315,11 @@ public class Data {
         this.coApplicants = coApplicants;
     }
 
-    public double getCommercialSpaceIncome() {
+    public Integer getCommercialSpaceIncome() {
         return commercialSpaceIncome;
     }
 
-    public void setCommercialSpaceIncome(double commercialSpaceIncome) {
+    public void setCommercialSpaceIncome(Integer commercialSpaceIncome) {
         this.commercialSpaceIncome = commercialSpaceIncome;
     }
 
@@ -403,27 +403,27 @@ public class Data {
         this.eTin = eTin;
     }
 
-    public double getEmiOfOtherLoan() {
+    public Integer getEmiOfOtherLoan() {
         return emiOfOtherLoan;
     }
 
-    public void setEmiOfOtherLoan(double emiOfOtherLoan) {
+    public void setEmiOfOtherLoan(Integer emiOfOtherLoan) {
         this.emiOfOtherLoan = emiOfOtherLoan;
     }
 
-    public double getFactoryIncome() {
+    public Integer getFactoryIncome() {
         return factoryIncome;
     }
 
-    public void setFactoryIncome(double factoryIncome) {
+    public void setFactoryIncome(Integer factoryIncome) {
         this.factoryIncome = factoryIncome;
     }
 
-    public double getFamilyExpenditure() {
+    public Integer getFamilyExpenditure() {
         return familyExpenditure;
     }
 
-    public void setFamilyExpenditure(double familyExpenditure) {
+    public void setFamilyExpenditure(Integer familyExpenditure) {
         this.familyExpenditure = familyExpenditure;
     }
 
@@ -443,11 +443,11 @@ public class Data {
         this.fee = fee;
     }
 
-    public double getInterestIncomeOfFDR() {
+    public Integer getInterestIncomeOfFDR() {
         return interestIncomeOfFDR;
     }
 
-    public void setInterestIncomeOfFDR(double interestIncomeOfFDR) {
+    public void setInterestIncomeOfFDR(Integer interestIncomeOfFDR) {
         this.interestIncomeOfFDR = interestIncomeOfFDR;
     }
 
@@ -467,11 +467,11 @@ public class Data {
         this.leadReferenceNo = leadReferenceNo;
     }
 
-    public double getLoanRequired() {
+    public Integer getLoanRequired() {
         return loanRequired;
     }
 
-    public void setLoanRequired(double loanRequired) {
+    public void setLoanRequired(Integer loanRequired) {
         this.loanRequired = loanRequired;
     }
 
@@ -539,11 +539,11 @@ public class Data {
         this.motherName = motherName;
     }
 
-    public double getNetSalary() {
+    public Integer getNetSalary() {
         return netSalary;
     }
 
-    public void setNetSalary(double netSalary) {
+    public void setNetSalary(Integer netSalary) {
         this.netSalary = netSalary;
     }
 
@@ -691,11 +691,11 @@ public class Data {
         this.relationshipWithApplicant = relationshipWithApplicant;
     }
 
-    public double getRemittanceIncome() {
+    public Integer getRemittanceIncome() {
         return remittanceIncome;
     }
 
-    public void setRemittanceIncome(double remittanceIncome) {
+    public void setRemittanceIncome(Integer remittanceIncome) {
         this.remittanceIncome = remittanceIncome;
     }
 
@@ -707,11 +707,11 @@ public class Data {
         this.rmCode = rmCode;
     }
 
-    public double getSecurityValue() {
+    public Integer getSecurityValue() {
         return securityValue;
     }
 
-    public void setSecurityValue(double securityValue) {
+    public void setSecurityValue(Integer securityValue) {
         this.securityValue = securityValue;
     }
 
@@ -723,11 +723,11 @@ public class Data {
         this.segment = segment;
     }
 
-    public double getSemipakaIncome() {
+    public Integer getSemipakaIncome() {
         return semipakaIncome;
     }
 
-    public void setSemipakaIncome(double semipakaIncome) {
+    public void setSemipakaIncome(Integer semipakaIncome) {
         this.semipakaIncome = semipakaIncome;
     }
 
@@ -747,11 +747,11 @@ public class Data {
         this.status = status;
     }
 
-    public double getTutionIncome() {
+    public Integer getTutionIncome() {
         return tutionIncome;
     }
 
-    public void setTutionIncome(double tutionIncome) {
+    public void setTutionIncome(Integer tutionIncome) {
         this.tutionIncome = tutionIncome;
     }
